+++ src/Network/Network.Test/ScenarioTests/VirtualNetworkTests.cs
@@ -148,9 +148,34 @@ namespace Commands.Network.Test.ScenarioTests
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.sdnnrp)]
+        public void TestVirtualNetworkCRUDFlowTimeout()
+        {
+            TestRunner.RunTestScript("Test-VirtualNetworkCRUD-FlowTimeout");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.sdnnrp)]
+        public void TestVirtualNetworkPeeringSyncCRUD()
+        {
+            TestRunner.RunTestScript("Test-SyncVirtualNetworkPeeringCRUD");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.sdnnrp)]
         public void TestVirtualNetworkInEdgeZone()
         {
             TestRunner.RunTestScript("Test-VirtualNetworkInEdgeZone");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.sdnnrp)]
+        public void TestVirtualNetworkEdgeZone()
+        {
+            TestRunner.RunTestScript("Test-VirtualNetworkEdgeZone");
         }
     }
 }